requirements.txt
chardet==5.2.0
numpy==2.2.2
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
wandb==0.19.1
matplotlib